403.04.28

+++ lib/views/product/product_view.dart
@@ -143,8 +143,8 @@ class ProductView extends StatelessWidget {
   }
 
   cartPrice() {
-    double price = double.parse(product.price!);
-    double regularPrice = double.parse(product.regularPrice!);
+    int price = int.parse(product.price!);
+    int regularPrice = int.parse(product.regularPrice!);
     int percent = 0;
     String toman = ' تومان';
     Color textColor = Colors.black87;
